feat(app): organize LabwareSetup to add map vs list view
closes RCORE-140 RCORE-141
