+++ app/src/assets/localization/en/protocol_setup.json
@@ -129,5 +129,8 @@
   "labware_name": "Labware Name",
   "volume": "Volume",
   "slot_location": "Slot {{slotName}}",
-  "slot_number": "Slot Number"
+  "slot_number": "Slot Number",
+  "module_slot_location": "Slot {{slotName}}, {{moduleName}}",
+  "initial_location": "Initial Location",
+  "secure_labware_instructions": "Secure labware instructions"
 }
